requirements.txt
telethon
pillow
apscheduler
pytz
